move company to separate class
