+++ app/core/admin.py
@@ -1,8 +1,9 @@
 # Register your model here.
 from django.contrib import admin
 
-from core.models import Vehicle, VehicleType, WorkshopCategory, Company, Tender, Route, Workshop, Station, TrackLimit, \
+from core.models import Vehicle, VehicleType, WorkshopCategory, Tender, Route, Workshop, Station, TrackLimit, \
     Criterion, TransportRequirement, Track, Line
+from core.models.company import Company
 
 admin.site.register(Vehicle)
 admin.site.register(Workshop)
